Finished sent information from server to client.
ClientController has some comments detailing the message sent from the
server thread.

The GoFishGame class now has a method that spits out the list of unique
pairs per player.

Player class is the middle man for the above, it does however send a 0
if there are no pairs for that player

Hand class now makes a sorted set of all unique cards in the inactive
cards.

Card class now has a comparator.