Fixed queryPlayer() so it...
Matches the sequence diagram and does the listed functions.
Changed the checkRemovePlayer to lostCards and added the functionality
described in the sequence diagram.
Card.getValue() is now public.
The ServerThread now has a 5th stage for when the game is over.
Made a stub in the ClientThread for ST stage 5.
The deck class can now take a variable amount of cards (in the case a
player needs to draw a new hand).
Made CardGame.getStartingHand() protected.