Changed Hand.findMatches to ...
findMatchesForUTF(List <Card>) so it is less ambiguous and it can find
unique cards in both parts of the hand.
It also gives the duplicity of cards for each value

Methods that use the findMatchesForUTF have been given the necessary
changes.

Player.getCardListForUTF now uses player.getPairs for collecting cards
into string format.

Server now sends to clients their active cards in the format of
##S
the first # is copies of cards with this value
The second # is the card value
The S is the Suit of the faceup card