Server to Client Information is now stored in the client
Fixed an issue where the host's client would not go to the proper
conditional branch when "PLAY" was sent out.

Client GUI now has references to a playerID (their spot in the following
arrays), a count of the cards in the deck, a count of cards from each
player, and a list of pairs each player has.

Fixed the Player.sendCardListForUTF so it sends all cards in the active
hand and not just the unique cards.

Fixed a couple built strings that didn't print their list and printed an
extra ';'.