Added initializer methods for a cardButtons[]...
and PlayerButtons[].
Added a "playa.png" to represent a player portrait.
All the above above buttons write a string to a new String[] userSelect
which holds the users selection for their turn.
!!!Started work on a method to update the list of buttons the player can
work with but it is not finished!!!
Added a numberToSymbol() and symbolToWord methods to get string
representations of the card int values and Symbols.
Added Final ints from Antonio's GUI Controller update branch for holding
a consistant card size.
Increased the size of the initial box to match the size listed in
Antonio's GUI Controller update branch


All the above is from ClientGUI



Card:
Added a getInt function to get the int value of its enums.

Hand:
Added a getUCards method to get a list of all the unique cards in the
users hand.

ClientContoller:
Server sends the names of all players to everyone. For use with the
playerButtons[]